+++ skrub/_reporting/tests/test_table_report.py
@@ -7,11 +7,27 @@ from pathlib import Path
 
 import numpy as np
 import pytest
+from sklearn.utils import Bunch
 
-from skrub import TableReport, ToDatetime, datasets
+from skrub import TableReport, ToDatetime
 from skrub import _dataframe as sbd
 from skrub._reporting._sample_table import make_table
 from skrub.conftest import polars_installed_without_pyarrow
+
+
+@pytest.fixture
+def simple_df(df_module):
+    return df_module.make_dataframe(
+        {
+            "A": [1, 2, 3, 4, 5],
+            "B": ["a", "b", "a", "b", "c"],
+        }
+    )
+
+
+@pytest.fixture
+def simple_series(df_module):
+    return df_module.make_column(name="A", values=[1, 2, 3, 4, 5])
 
 
 def get_report_id(html):
@@ -294,8 +310,8 @@ def test_minimal_mode(pd_module):
     assert 'id="column-associations-panel"' not in html
 
 
-def test_error_input_type():
-    df = datasets.fetch_employee_salaries()
+def test_error_input_type(simple_df, simple_series):
+    df = Bunch(X=simple_df, y=simple_series)
     with pytest.raises(TypeError):
         TableReport(df)